feat: add release note generator script
This commit adds a new bash script to generate a release note for the current release. The script uses git diff from previous tag to HEAD and echos it all out to stdout. 

The `ci.yml` file was also updated with a new step that runs the `release_note_generator.sh` script and saves its output in `release_note.md`. This file is then used as the body of the commit message generated by ncipollo/release-action.

The purpose of this change is to automate the process of generating release notes, making it easier for developers and users alike to understand what has changed in each version.