feat: add release workflow
Add a new workflow to build and release binary files. The workflow is triggered by push or pull request events on the `release` branch. The job consists of building the binaries using Nuitka, uploading them as artifacts, generating a release note, and creating a new tag with the version number.

The commit also includes changes to the existing CI workflow where it now ignores push events on the `release` branch.