feat: add version tag to release.yml
Add a new feature that adds a version tag to the release.yml file. The script now reads the pyproject.toml file and extracts the current version number, which is then used in creating a new tag name with format "v${VERSION}-${{ github.run_number }}-${{ github.sha }}". This allows for easier tracking of releases and better organization of code changes.