feat: update artifact upload action to v3
Update the `upload-artifact` action from version 2 to version 3 in the release workflow. This change was made to improve performance and reliability of uploading artifacts for each operating system (Ubuntu, Windows, MacOS). The new version will be used in all three jobs that upload different artifacts depending on the OS being used.