feat: add release workflow
This commit adds a GitHub Actions workflow for releasing new versions of the project. The `release.yml` file has been modified to include environment variables that specify the name and tag of the release, as well as a release note. These variables are used by the `ncipollo/release-action` action to create a new GitHub Release with appropriate metadata.

The changes made in this commit allow for easier and more consistent releases, which should help streamline development workflows going forward.