feat(ci_worker): remove ci_worker.py
Remove the `ci_worker.py` file that contained a helper to multiplexing if the CI matrix environment. The script was used to execute different scripts based on the OS and action. It had two actions, "lint" and "build". However, it is no longer needed as we have moved our CI/CD pipeline to GitHub Actions.