build: remove unnecessary condition for creating release tag
The if statement checking the branch name was removed from the job that creates a release tag. This is because it's not necessary to check for a specific branch since this job only runs when a new release is created. The email and user name used by git were also updated to match our CI configuration.