feat: Add binary files for Windows, Linux, and Mac (#2)
* build: remove unnecessary condition for creating release tag

The if statement checking the branch name was removed from the job that creates a release tag. This is because it's not necessary to check for a specific branch since this job only runs when a new release is created. The email and user name used by git were also updated to match our CI configuration.

* feat: Add binary files for Windows, Linux, and Mac

This commit adds the binary files for Windows (.exe), Linux (.elf), and Mac (.macho) to the latest release. The binaries can be linked to `/usr/local/bin/cg` on Unix-based systems using a command provided in this commit message.

The raw python code can still be used by following the instructions in README.md. Additionally, this commit marks two items as complete in Todo list - smarter prompt for commit message that uses score to measure quality of generated messages and tests with CI.

Only changes made are addition of binary files and updating Todo list.