feat: add version tag to release.yml (#1) (#3)
* feat: add version tag to release.yml

Add a new feature that adds a version tag to the release.yml file. The script now reads the pyproject.toml file and extracts the current version number, which is then used in creating a new tag name with format "v${VERSION}-${{ github.run_number }}-${{ github.sha }}". This allows for easier tracking of releases and better organization of code changes.

* feat: update artifact upload action to v3

Update the `upload-artifact` action from version 2 to version 3 in the release workflow. This change was made to improve performance and reliability of uploading artifacts for each operating system (Ubuntu, Windows, MacOS). The new version will be used in all three jobs that upload different artifacts depending on the OS being used.

* feat: update release workflow

This commit updates the release workflow to use the TAG_NAME variable instead of github.ref for naming and tagging releases. The TAG_NAME is read from tag_name.txt file. This change was made to ensure consistency in naming conventions across different workflows.

The `name` and `tag` parameters were updated in ncipollo/release-action@v1 action with appropriate values. The commit parameter now uses ${{ github.sha }} as before.

Also, added TAG_NAME environment variable with value from tag_name.txt file for use in jobs that require it.

* feat: add release workflow

This commit adds a GitHub Actions workflow for releasing new versions of the project. The `release.yml` file has been modified to include environment variables that specify the name and tag of the release, as well as a release note. These variables are used by the `ncipollo/release-action` action to create a new GitHub Release with appropriate metadata.

The changes made in this commit allow for easier and more consistent releases, which should help streamline development workflows going forward.

* feat(ci_worker): remove ci_worker.py

Remove the `ci_worker.py` file that contained a helper to multiplexing if the CI matrix environment. The script was used to execute different scripts based on the OS and action. It had two actions, "lint" and "build". However, it is no longer needed as we have moved our CI/CD pipeline to GitHub Actions.

* build: remove unnecessary condition for creating release tag

The if statement checking the branch name was removed from the job that creates a release tag. This is because it's not necessary to check for a specific branch since this job only runs when a new release is created. The email and user name used by git were also updated to match our CI configuration.

* feat: Add binary files for Windows, Linux, and Mac (#2)

* build: remove unnecessary condition for creating release tag

The if statement checking the branch name was removed from the job that creates a release tag. This is because it's not necessary to check for a specific branch since this job only runs when a new release is created. The email and user name used by git were also updated to match our CI configuration.

* feat: Add binary files for Windows, Linux, and Mac

This commit adds the binary files for Windows (.exe), Linux (.elf), and Mac (.macho) to the latest release. The binaries can be linked to `/usr/local/bin/cg` on Unix-based systems using a command provided in this commit message.

The raw python code can still be used by following the instructions in README.md. Additionally, this commit marks two items as complete in Todo list - smarter prompt for commit message that uses score to measure quality of generated messages and tests with CI.

Only changes made are addition of binary files and updating Todo list.

* feat: update release workflow

This commit updates the release workflow to use actions/download-artifact@v3 instead of v2. It also generates environment variables for tag name, release note and release name using the output of `release_note_generator.sh`. This will make it easier to create a new GitHub Release with accurate information.